+++ java-interop/x10.runtime/src-java/x10/core/GlobalRef.java
@@ -30,7 +30,7 @@ import x10.x10rt.X10JavaSerializer;
 public final class GlobalRef<T> extends x10.core.Struct implements
         Externalizable, X10JavaSerializable {
 	
-	private static final int _serialization_id = x10.x10rt.DeserializationDispatcher.addDispatcher(GlobalRef.class, "x10.lang.GlobalRef");
+	private static final short _serialization_id = x10.x10rt.DeserializationDispatcher.addDispatcher(GlobalRef.class, "x10.lang.GlobalRef");
 
     public static final x10.rtt.RuntimeType<GlobalRef<?>> $RTT = new x10.rtt.NamedType<GlobalRef<?>>(
         "x10.lang.GlobalRef",
@@ -312,7 +312,7 @@ public final class GlobalRef<T> extends x10.core.Struct implements
         return $_deserialize_body(gr, deserializer);
 	}
 
-	public int $_get_serialization_id() {
+	public short $_get_serialization_id() {
 		return _serialization_id;
 	}
 
@@ -340,7 +340,7 @@ public final class GlobalRef<T> extends x10.core.Struct implements
     public static class LocalEval extends x10.core.Ref {
 
 	private static final long serialVersionUID = 1L;
-    private static final int _serialization_id = x10.x10rt.DeserializationDispatcher.addDispatcher(LocalEval.class);
+    private static final short _serialization_id = x10.x10rt.DeserializationDispatcher.addDispatcher(LocalEval.class);
 	public static final x10.rtt.RuntimeType<LocalEval> $RTT = new x10.rtt.NamedType<LocalEval>("x10.lang.GlobalRef.LocalEval", LocalEval.class, new x10.rtt.Type[] {x10.rtt.Types.OBJECT});
 	public x10.rtt.RuntimeType<?> $getRTT() {return $RTT;}
     
@@ -377,7 +377,7 @@ public final class GlobalRef<T> extends x10.core.Struct implements
         return $_deserialize_body(obj, deserializer);
 	}
 
-	public int $_get_serialization_id() {
+	public short $_get_serialization_id() {
 		return _serialization_id;
 	}
 
@@ -388,7 +388,7 @@ public final class GlobalRef<T> extends x10.core.Struct implements
 
 	public static class $Closure$Eval<$T, $U> extends x10.core.Ref implements x10.core.fun.Fun_0_0 {
 	    private static final long serialVersionUID = 1L;
-        private static final int _serialization_id = x10.x10rt.DeserializationDispatcher.addDispatcher($Closure$Eval.class);
+        private static final short _serialization_id = x10.x10rt.DeserializationDispatcher.addDispatcher($Closure$Eval.class);
 	    public static final x10.rtt.RuntimeType<$Closure$Eval> $RTT =
 		new x10.rtt.StaticFunType<$Closure$Eval>($Closure$Eval.class, 
 							 new x10.rtt.RuntimeType.Variance[] {x10.rtt.RuntimeType.Variance.INVARIANT, x10.rtt.RuntimeType.Variance.INVARIANT},
@@ -432,7 +432,7 @@ public final class GlobalRef<T> extends x10.core.Struct implements
             return $_deserialize_body(obj, deserializer);
         }
 
-        public int $_get_serialization_id() {
+        public short $_get_serialization_id() {
             return _serialization_id;
         }
 
@@ -452,7 +452,7 @@ public final class GlobalRef<T> extends x10.core.Struct implements
             
 	public static class $Closure$Apply<$T> extends x10.core.Ref implements x10.core.fun.Fun_0_0 {
 	    private static final long serialVersionUID = 1L;
-        private static final int _serialization_id = x10.x10rt.DeserializationDispatcher.addDispatcher($Closure$Apply.class);
+        private static final short _serialization_id = x10.x10rt.DeserializationDispatcher.addDispatcher($Closure$Apply.class);
 	    public static final x10.rtt.RuntimeType<$Closure$Apply> $RTT =
 		new x10.rtt.StaticFunType<$Closure$Apply>($Closure$Apply.class, 
 							  new x10.rtt.RuntimeType.Variance[] {x10.rtt.RuntimeType.Variance.INVARIANT},
@@ -490,7 +490,7 @@ public final class GlobalRef<T> extends x10.core.Struct implements
             return $_deserialize_body(obj, deserializer);
         }
 
-        public int $_get_serialization_id() {
+        public short $_get_serialization_id() {
             return _serialization_id;
         }
