+++ java-interop/x10.runtime/src-java/x10/core/concurrent/AtomicReference.java
@@ -25,7 +25,7 @@ import java.io.IOException;
 public final class AtomicReference<T> extends java.util.concurrent.atomic.AtomicReference<T> implements RefI, X10JavaSerializable {
 
     private static final long serialVersionUID = 1L;
-    private static final int _serialization_id = x10.x10rt.DeserializationDispatcher.addDispatcher(AtomicReference.class);
+    private static final short _serialization_id = x10.x10rt.DeserializationDispatcher.addDispatcher(AtomicReference.class);
 
     // constructor just for allocation
     public AtomicReference(java.lang.System[] $dummy) {
@@ -74,7 +74,7 @@ public final class AtomicReference<T> extends java.util.concurrent.atomic.Atomic
 		serializer.write(get());
 	}
 
-	public int $_get_serialization_id() {
+	public short $_get_serialization_id() {
 		return _serialization_id;
 	}
